changed error in Stereopsis.java, mask for rightEye was still short, so comparison with address did not work
git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@520 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/caviar/stereopsis/Stereopsis.java
@@ -25,7 +25,7 @@ public class Stereopsis {
     static final int bit=15;
     
     /** the bit mask used for constructing a right eye event from a raw event */
-    public static  final short MASK_RIGHT_ADDR=(short)(1<<bit);
+    public static  final int MASK_RIGHT_ADDR=(int)(1<<bit);
     
     
     /** @return true if raw address signals left eye event */
